refactor #choose_cell && refactor #make_new_choice
removed call to #choose_cell from #make_new_choice, and removed parameters as
they were no longer needed because of this change. added the new choice as
the return value.

added control flow to #choose_cell to account for:
  1) selected numbers outside of the board (anything not in 1..9)
  2) an already-chosen space being selected again.

if either of these conditions are truthy, then a call to #make_new_choice
is made to select another square, and that new value is stored in new_choice.
a new call to #choose_cell is returned with the current marker ('X' or 'O')
and the value of new_choice.